Bump contexts properties to EarthObservationExperiment API
