+++ src/eradiate/experiments/_atmosphere.py
@@ -171,40 +171,6 @@ class AtmosphereExperiment(EarthObservationExperiment):
         return kwargs
 
     @property
-    def contexts(self) -> list[KernelDictContext]:
-        # Inherit docstring
-
-        # Collect contexts from all measures
-        sctxs = []
-
-        for measure in self.measures:
-            sctxs.extend(measure.spectral_cfg.spectral_ctxs())
-
-        # Sort and remove duplicates
-        key = {
-            MonoSpectralContext: lambda sctx: sctx.wavelength.m,
-            CKDSpectralContext: lambda sctx: (
-                sctx.bindex.bin.wcenter.m,
-                sctx.bindex.index,
-            ),
-        }[type(sctxs[0])]
-
-        sctxs = deduplicate_sorted(
-            sorted(sctxs, key=key), cmp=lambda x, y: key(x) == key(y)
-        )
-        kwargs = self._context_kwargs
-
-        return [KernelDictContext(spectral_ctx=sctx, kwargs=kwargs) for sctx in sctxs]
-
-    @property
-    def context_init(self) -> KernelDictContext:
-        # Inherit docstring
-
-        return KernelDictContext(
-            spectral_ctx=SpectralContext.new(), kwargs=self._context_kwargs
-        )
-
-    @property
     def scene_objects(self) -> dict[str, SceneElement]:
         # Inherit docstring
